bug #22 TimePicker in Grid causes TypeError
the problem is the FIXME lines in timepickerConnector.js
 - they access the shadow DOM, but it hasn't been inititalized
   when setting the locale upon initial attachment.

This solution just makes a dumy JS void(0) call, and upon success,
asynchronously calls setLocale().  This allows the client to fully
initialize the element and nodes before attempting to directly acccess
them in an unsafe manner.

I'm sure there is a better way, thus the "FIXME" comment in there, but
as this is my first foray into Vaadin 10+, I'll stick to the server side.

This can be verified as correct by inserting this before the sample code in the issue:

UI.getCurrent().setLocale(Locale.CHINA);

and observing that the time is displayed with the proper format upon first render in the UI.