bug #22 fix bad local merge
Had the old and new code both in there by mistake.

Also reworked the fix to use a PendingJavaScriptResult, to redo the call
to setLocale() only if it initially fails (which it will until the order
of operations bug in the client side code is fixed).

This version has been tested with the sample code from the original
issue, and properly displays a localized time inside a grid on initial
render.