bug #22 Took a stab at a unit test
It definitely displays the errors mentioned in the issue comments when
run with the current version of the component, and correctly displays
the component and formats the value on initial attachment when run with
the branch code.  I can't run the integration tests locally, though, so
we will see how they do in the CI environment.

Tests are based on the localization tests, but here use a page that
always initializes the component and UI with the given locale, rather
than reusing the component, to simulate first display of a component
with a value in a session with a selected locale.